+++ src/Commands/InvalidateCacheCommand.php
@@ -3,10 +3,12 @@
 namespace Rapidez\Statamic\Commands;
 
 use Illuminate\Console\Command;
-use Illuminate\Support\Facades\Cache;
+use Illuminate\Contracts\Filesystem\FileNotFoundException;
 use Illuminate\Support\Facades\DB;
 use Rapidez\Core\Facades\Rapidez;
+use Statamic\Facades\File;
 use Statamic\StaticCaching\Cacher;
+use Statamic\StaticCaching\Cachers\Writer;
 
 class InvalidateCacheCommand extends Command
 {
@@ -18,18 +20,18 @@ class InvalidateCacheCommand extends Command
 
     public $latestCheck;
 
-    public function handle(Cacher $cacher): void
+    public function handle(Cacher $cacher, Writer $writer): void
     {
-        $this->latestCheck = Cache::get($this->signature);
+        $this->latestCheck = $this->getLatestCheckDate();
 
         if (!$this->latestCheck) {
             $this->info('Cleared all urls (as we do not have a latest check date yet)');
-            $this->setLatestCheckDate();
             $cacher->flush();
+            $this->setLatestCheckDate($writer);
             return;
         }
+        $this->setLatestCheckDate($writer);
 
-        $this->setLatestCheckDate();
         $stores = Rapidez::getStores();
 
         foreach ($stores as $store) {
@@ -57,6 +59,7 @@ class InvalidateCacheCommand extends Command
     {
         $products = config('rapidez.models.product')::withoutGlobalScopes()
             ->where('updated_at', '>=', $this->latestCheck)
+            ->orWhereIn('entity_id', $this->getUpdatedStockProducts())
             ->with(['parent:entity_id' => ['rewrites']])
             ->with('rewrites')
             ->get('entity_id');
@@ -70,6 +73,26 @@ class InvalidateCacheCommand extends Command
         }
 
         return $this;
+    }
+
+    protected function getUpdatedStockProducts()
+    {
+        $currentStock = DB::table('cataloginventory_stock_item')
+            ->pluck('qty', 'product_id')->toArray();
+
+        $previousStock = $this->getPreviousStock();
+
+        $this->setPreviousStock($currentStock);
+
+        return array_keys(
+            array_diff_assoc(
+                (array) $currentStock,
+                (array) $previousStock
+            ) + array_diff_assoc(
+                (array) $previousStock,
+                (array) $currentStock
+            )
+        );
     }
 
     protected function addCategoryUrls(): self
@@ -115,12 +138,40 @@ class InvalidateCacheCommand extends Command
         return $this;
     }
 
-    protected function setLatestCheckDate(): void
+    protected function getLatestCheckDate()
     {
-        // With this we're just making sure the comparison
-        // is done within the same timezone in MySQL.
-        $now = DB::select('SELECT NOW() AS `current_time`');
+        try {
+            return File::get(config('statamic.static_caching.strategies.full.path') . '/.last-invalidation');
+        } catch (FileNotFoundException $e) {
+            return null;
+        }
+    }
 
-        Cache::forever($this->signature, $now[0]->current_time);
+    protected function setLatestCheckDate(Writer $writer): void
+    {
+        $writer->write(
+            config('statamic.static_caching.strategies.full.path') . '/.last-invalidation',
+            // With this we're just making sure the comparison
+            // is done within the same timezone in MySQL.
+            DB::selectOne('SELECT NOW() AS `current_time`')->current_time
+        );
+    }
+
+    protected function getPreviousStock()
+    {
+        try {
+            return json_decode(File::get(config('statamic.static_caching.strategies.full.path') . '/.product-stocks'), true, 512, JSON_THROW_ON_ERROR);
+        } catch (FileNotFoundException|\JsonException $e) {
+            return [];
+        }
+    }
+
+    protected function setPreviousStock(array $previousStock): void
+    {
+        $writer = app(Writer::class);
+        $writer->write(
+            config('statamic.static_caching.strategies.full.path') . '/.product-stocks',
+            json_encode($previousStock)
+        );
     }
 }